refactor(edge-worker): Update project configuration and remove unused function files
Clean up Supabase edge worker project structure by removing deprecated function files, updating
import paths, and simplifying project configuration

- Remove unused Supabase function files (creating_queue, failing_always, increment_sequence,
serial_sleep)
- Update import paths in remaining function files to use distributed index
- Modify project.json configuration and dependencies
- Remove unnecessary .gitkeep and update deno.lock file
