+++ pkgs/edge-worker/deno.lock
@@ -4,6 +4,7 @@
     "specifiers": {
       "jsr:@deno-library/progress": "jsr:@deno-library/progress@1.5.1",
       "jsr:@henrygd/queue@^1.0.7": "jsr:@henrygd/queue@1.0.7",
+      "jsr:@std/assert": "jsr:@std/assert@0.224.0",
       "jsr:@std/assert@^0.224.0": "jsr:@std/assert@0.224.0",
       "jsr:@std/async@^0.224.0": "jsr:@std/async@0.224.2",
       "jsr:@std/fmt@1.0.3": "jsr:@std/fmt@1.0.3",
@@ -12,8 +13,6 @@
       "jsr:@std/io@0.225.0": "jsr:@std/io@0.225.0",
       "jsr:@std/io@^0.225.0": "jsr:@std/io@0.225.0",
       "npm:@henrygd/queue@^1.0.7": "npm:@henrygd/queue@1.0.7",
-      "npm:@pgflow/core@^0.0.10": "npm:@pgflow/core@0.0.10",
-      "npm:@pgflow/dsl@^0.0.10": "npm:@pgflow/dsl@0.0.10",
       "npm:pino@^9.6.0": "npm:pino@9.6.0",
       "npm:postgres@3.4.5": "npm:postgres@3.4.5"
     },
@@ -57,17 +56,6 @@
     "npm": {
       "@henrygd/queue@1.0.7": {
         "integrity": "sha512-Jmt/iO6yDlz9UYGILkm/Qzi/ckkEiTNZcqDvt3QFLE4OThPeiCj6tKsynHFm/ppl8RumWXAx1dZPBPiRPaaGig==",
-        "dependencies": {}
-      },
-      "@pgflow/core@0.0.10": {
-        "integrity": "sha512-MUYSM7zAtkpGI31JlQehCeE1gA7PVm2L9e4rfCAnNw/Do0LJaQvKrHRGtlb7gK6tgOwdAZcsQGWH+PW7UbRPDA==",
-        "dependencies": {
-          "@pgflow/dsl": "@pgflow/dsl@0.0.10",
-          "postgres": "postgres@3.4.5"
-        }
-      },
-      "@pgflow/dsl@0.0.10": {
-        "integrity": "sha512-n1rQ7vk5yD5uYAt6ukxasLT4a8vv+hLIoKXW7xOSQSRQN4xvdWMdClDBkgl3UpaIJaFUoK2nzYzF26jqoVqYqw==",
         "dependencies": {}
       },
       "atomic-sleep@1.0.0": {
@@ -439,8 +427,6 @@
     "packageJson": {
       "dependencies": [
         "npm:@henrygd/queue@^1.0.7",
-        "npm:@pgflow/core@^0.0.10",
-        "npm:@pgflow/dsl@^0.0.10",
         "npm:pino@^9.6.0",
         "npm:postgres@3.4.5"
       ]
