--- pkgs/edge-worker/supabase/functions/creating_queue/index.ts
@@ -1,5 +1,0 @@
-import { EdgeWorker } from '../_src/EdgeWorker.ts';
-
-EdgeWorker.start(console.log, {
-  queueName: 'creating_queue',
-});
