--- pkgs/edge-worker/supabase/functions/failing_always/index.ts
@@ -1,13 +1,0 @@
-import { EdgeWorker } from '../_src/EdgeWorker.ts';
-
-function failingAlways() {
-  console.log('(╯°□°)╯︵ ┻━┻');
-  throw new Error('(╯°□°)╯︵ ┻━┻');
-}
-
-EdgeWorker.start(failingAlways, {
-  queueName: 'failing_always',
-  retryLimit: 2,
-  retryDelay: 2,
-  maxPollSeconds: 1,
-});
